fix: UserCharacterRepo id Integer로 받도록 변경

+++ src/main/java/com/luckkids/domain/userCharacter/UserCharacterRepository.java
@@ -2,5 +2,5 @@ package com.luckkids.domain.userCharacter;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserCharacterRepository extends JpaRepository<UserCharacter, Long> {
+public interface UserCharacterRepository extends JpaRepository<UserCharacter, Integer> {
 }
